+++ examples/Order/Transaction/Delete.php
@@ -9,17 +9,19 @@ require_once '../../../vendor/autoload.php';
 
 use MercadoPago\Client\Common\RequestOptions;
 use MercadoPago\Client\Order\OrderClient;
+use MercadoPago\Client\Order\OrderTransactionClient;
 use MercadoPago\Exceptions\MPApiException;
 use MercadoPago\MercadoPagoConfig;
 
 // Step 2: Set production or sandbox access token
-MercadoPagoConfig::setAccessToken("APP_USR-874202490252970-100714-e890db6519b0dceb4ef24ef41ed816e4-2021490138");
+MercadoPagoConfig::setAccessToken("<ACCESS_TOKEN>");
 // Step 2.1 (optional - default is SERVER): Set your runtime enviroment from MercadoPagoConfig::RUNTIME_ENVIROMENTS
 // In case you want to test in your local machine first, set runtime enviroment to LOCAL
 MercadoPagoConfig::setRuntimeEnviroment(MercadoPagoConfig::LOCAL);
 
-// Step 3: Initialize the API client
+// Step 3: Initialize the API client for Order and Transaction
 $client = new OrderClient();
+$client_transactions = new OrderTransactionClient();
 
 try {
     // Step 4: Create the request array
@@ -38,7 +40,7 @@ try {
                   "payment_method" => [
                       "id" => "master",
                       "type" => "credit_card",
-                      "token" => "699419faca6def0fdc98743f4e9f72f8",
+                      "token" => "<CARD_TOKEN>",
                       "installments" => 1,
                   ]
               ]
@@ -46,13 +48,13 @@ try {
       ],
 ];
     $request_options = new RequestOptions();
-    $request_options->setCustomHeaders(["X-Idempotency-Key: 141221198ygh77234432", "X-Sandbox: true"]);
+    $request_options->setCustomHeaders(["X-Idempotency-Key:<SOME_UNIQUE_VALUE>"]);
     $order = $client->create($request, $request_options);
     $transaction_id = $order->transactions->payments[0]->id;
     $order_id = $order->id;
 
-    $request_options->setCustomHeaders(["X-Idempotency-Key: 141221116609876567", "X-Sandbox: true"]);
-    $response = $client->deleteTransaction($order_id, $transaction_id, $request_options);
+    $request_options->setCustomHeaders(["X-Idempotency-Key:<SOME_UNIQUE_VALUE>"]);
+    $response = $client_transactions->deleteTransaction($order_id, $transaction_id, $request_options);
     if ($response === null || $response->getStatusCode() === 204) {
         echo "Transaction deleted successfully. HTTP Status Code: 204\n";
     } else {
